Extract Play API tests into separate file
